refactor task types to be in separate package
